# api-contract.yaml
openapi: 3.0.3
info:
  title: CODELESS API
  version: 0.6.0
  description: >
    API contract for the CODELESS multi-model AI chat platform.
    Features: authentication (sign-up/login/logout), user profile with
    custom avatar and preferences, multi-turn chat with AI models
    (Server-Sent Events streaming & JSON fallback), model switching
    mid-conversation, admin management of PROVIDERS (CRUD, active toggle,
    per-provider default), admin management of MODELS (CRUD, active toggle,
    platform default, provider binding), admin global token quota limit
    per user, customizable system prompt, admin user listing & management
    with chat/token stats, admin dashboard KPI metrics, forwarding chat to
    real OpenAI-compatible upstreams, and standard OpenAI-compatible endpoints
    (/v1/models, /v1/chat/completions) behind bearer auth.
    All standard JSON responses are wrapped in a unified envelope:
    { success: boolean, message: string, data: any }.
  contact:
    name: CODELESS Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development API base
  - url: https://api.codeless.example.com/api/v1
    description: Production API base (placeholder)
...
security:
  - bearerAuth: []
paths:
...
  /admin/models:
    get:
      tags: [Admin - Models]
      summary: List all configured AI models
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelsListApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Admin - Models]
      summary: Add a new AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelRequest'
      responses:
        '201':
          description: Model added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/models/{modelId}:
    parameters:
      - $ref: '#/components/parameters/ModelId'
    patch:
      tags: [Admin - Models]
      summary: Update model properties (name, provider, providerId, apiIdentifier, apiKey, baseUrl, isActive)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModelRequest'
      responses:
        '200':
          description: Model updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Admin - Models]
      summary: Remove an AI model
      responses:
        '204':
          description: Model removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/models/{modelId}/default:
    parameters:
      - $ref: '#/components/parameters/ModelId'
    patch:
      tags: [Admin - Models]
      summary: Set this model as the platform-wide default
      description: >
        Kept here because chat needs a fallback model when a conversation
        doesn't specify one (see the original Day 1-2 "set default model"
        requirement) — remove this endpoint if that's out of scope for you.
      responses:
        '200':
          description: Model set as default
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/models/{modelId}/status:
    parameters:
      - $ref: '#/components/parameters/ModelId'
    patch:
      tags: [Admin - Models]
      summary: Toggle active/inactive status of an AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModelStatusRequest'
      responses:
        '200':
          description: Model status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelApiResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ModelId:
      name: modelId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not allowed to perform this action (admin only)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    UpdateModelRequest:
      type: object
      properties:
        name:
          type: string
        provider:
          type: string
        providerId:
          type: string
          format: uuid
        apiIdentifier:
          type: string
        apiKey:
          type: string
        baseUrl:
          type: string
        isActive:
          type: boolean
...
    CreateModelRequest:
      type: object
      required: [name, provider, apiIdentifier]
      properties:
        name:
          type: string
          description: Display name shown in the admin panel and chat UI.
        provider:
          type: string
          description: >
            Provider name. Auto-registers the provider if the name is new
            (backward compatible); prefer providerId when available.
        providerId:
          type: string
          format: uuid
          description: Optional — link the model to an existing provider row.
        apiIdentifier:
          type: string
          description: The model name/id sent to the provider's API.
        apiKey:
          type: string
          description: Optional provider API key (masked on read).
        baseUrl:
          type: string
          description: Optional custom base URL for OpenAI-compatible LLM endpoints.
        isActive:
          type: boolean
          default: true
    UpdateModelStatusRequest:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
          description: Whether the model is active and selectable for chat.
    Model:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        provider:
          type: string
        providerId:
          type: string
          format: uuid
          nullable: true
          description: Owning provider row (models cascade-delete with it).
        apiIdentifier:
          type: string
        apiKey:
          type: string
          description: Masked provider API key.
          nullable: true
        baseUrl:
          type: string
          description: Custom base URL.
          nullable: true
        isActive:
          type: boolean
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Error:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Invalid email or password"
        data:
          nullable: true
          example: null
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
    ApiResponse:
      type: object
      required: [success, message, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          nullable: true
          description: Response payload
...
    ModelApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Model'
    ModelsListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Model'
...